move history to class

+++ esphome/components/airquality/abstract_aqi_calculator.h
@@ -8,7 +8,7 @@ enum Pollutant : uint8_t { PM25 = 0, PM10, NO2, O3, CO, SO2 };
 
 class AbstractAQICalculator {
  public:
-  virtual uint8_t calculate_index(uint16_t value, Pollutant pollutant_type) = 0;
+  virtual uint16_t calculate_index(uint16_t value, Pollutant pollutant_type) = 0;
 };
 
 }  // namespace airquality
